+++ apps/backend/src/modules/activity-logs/libs/types/types.ts
@@ -1,6 +1,9 @@
 export {
 	type ActivityLogCreateItemResponseDto,
 	type ActivityLogCreateRequestDto,
+	type ActivityLogGetAllAnalyticsResponseDto,
+	type ActivityLogGetAllItemAnalyticsResponseDto,
 	type ActivityLogGetAllItemResponseDto,
 	type ActivityLogGetAllResponseDto,
+	type ActivityLogQueryParameters,
 } from "@git-fit/shared";