+++ apps/backend/src/modules/project-groups/project-group.controller.ts
@@ -70,7 +70,12 @@ class ProjectGroupController extends BaseController {
 				),
 			method: "POST",
 			path: ProjectGroupsApiPath.ROOT,
-			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_USER_ACCESS])],
+			preHandlers: [
+				checkUserPermissions([
+					PermissionKey.MANAGE_USER_ACCESS,
+					PermissionKey.MANAGE_ALL_PROJECTS,
+				]),
+			],
 			validation: {
 				body: projectGroupCreateValidationSchema,
 			},
@@ -81,7 +86,12 @@ class ProjectGroupController extends BaseController {
 				this.delete(options as APIHandlerOptions<{ params: { id: string } }>),
 			method: "DELETE",
 			path: ProjectGroupsApiPath.$ID,
-			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_USER_ACCESS])],
+			preHandlers: [
+				checkUserPermissions([
+					PermissionKey.MANAGE_USER_ACCESS,
+					PermissionKey.MANAGE_ALL_PROJECTS,
+				]),
+			],
 		});
 
 		this.addRoute({
@@ -94,7 +104,12 @@ class ProjectGroupController extends BaseController {
 				),
 			method: "GET",
 			path: ProjectGroupsApiPath.$ID,
-			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_USER_ACCESS])],
+			preHandlers: [
+				checkUserPermissions([
+					PermissionKey.MANAGE_USER_ACCESS,
+					PermissionKey.MANAGE_ALL_PROJECTS,
+				]),
+			],
 		});
 
 		this.addRoute({
@@ -107,6 +122,12 @@ class ProjectGroupController extends BaseController {
 				),
 			method: "PATCH",
 			path: ProjectGroupsApiPath.$ID,
+			preHandlers: [
+				checkUserPermissions([
+					PermissionKey.MANAGE_USER_ACCESS,
+					PermissionKey.MANAGE_ALL_PROJECTS,
+				]),
+			],
 			validation: {
 				body: projectGroupPatchValidationSchema,
 			},