+++ apps/frontend/src/pages/analytics/analytics.tsx
@@ -1,47 +1,83 @@
-import { DateInput, PageLayout, Select } from "~/libs/components/components.js";
+import { DateInput, Loader, PageLayout, Select } from "~/libs/components/components.js";
+import { DataStatus } from "~/libs/enums/enums.js";
 import { subtractDays } from "~/libs/helpers/helpers.js";
 import {
-	//useAppDispatch,
+	useAppDispatch,
 	useAppForm,
 	useAppSelector,
 	useCallback,
-	//	useEffect,
+	useEffect,
+	useFormWatch,
 } from "~/libs/hooks/hooks.js";
+import { actions as activityLogActions } from "~/modules/activity/activity.js";
 
+import { AnalyticsTable } from "./libs/components/components.js";
 import { ANALYTICS_DATE_MAX_RANGE } from "./libs/constants/constants.js";
 import { getProjectOptions } from "./libs/helpers/helpers.js";
 import styles from "./styles.module.css";
 //import { actions as projectActions } from "~/modules/projects/projects.js";
 
 const Analytics = (): JSX.Element => {
+	const dispatch = useAppDispatch();
 	const todayDate = new Date();
 	//const dispatch = useAppDispatch();
 
-	// useEffect(() => {
-	// 	void dispatch(projectActions.loadAll());
-	// }, [dispatch]);
+	const { activityLogs, dataStatus } = useAppSelector(
+		({ activityLogs }) => activityLogs,
+	);
 
-	const { control, handleSubmit } = useAppForm<{
-		dateRange: [Date, Date];
-		projects: number;
-	}>({
+	const { control, handleSubmit, isDirty } = useAppForm({
 		defaultValues: {
-			dateRange: [subtractDays(todayDate, ANALYTICS_DATE_MAX_RANGE), todayDate],
+			dateRange: [
+				subtractDays(todayDate, ANALYTICS_DATE_MAX_RANGE),
+				todayDate,
+			] as [Date, Date],
 			projects: 0,
 		},
 	});
 	const { projects } = useAppSelector(({ projects }) => projects);
 
-	const handleFormSubmit = useCallback(
-		(event_: React.BaseSyntheticEvent): void => {
-			void handleSubmit(() => {
-				// TODO: implement this function
-			})(event_);
+	const dateRangeValue = useFormWatch({ control, name: "dateRange" });
+
+	const handleLoadLogs = useCallback(
+		([startDate, endDate]: [Date, Date]) => {
+			const formattedStartDate = startDate.toISOString();
+			const formattedEndDate = endDate.toISOString();
+
+			void dispatch(
+				activityLogActions.loadAll({
+					endDate: formattedEndDate,
+					startDate: formattedStartDate,
+				}),
+			);
 		},
-		[handleSubmit],
+		[dispatch],
 	);
 
+	useEffect(() => {
+		handleLoadLogs(dateRangeValue);
+	}, [dateRangeValue, handleLoadLogs]);
+
+	const handleFormSubmit = useCallback(
+		(event_?: React.BaseSyntheticEvent): void => {
+			void handleSubmit((formData) => {
+				handleLoadLogs(formData.dateRange);
+			})(event_);
+		},
+		[handleLoadLogs, handleSubmit],
+	);
+
+
 	const projectOptions = getProjectOptions(projects);
+
+	useEffect(() => {
+		if (isDirty) {
+			handleFormSubmit();
+		}
+	}, [dateRangeValue, isDirty, handleFormSubmit]);
+
+	const isLoading =
+		dataStatus === DataStatus.IDLE || dataStatus === DataStatus.PENDING;
 
 	return (
 		<PageLayout>
@@ -67,6 +103,14 @@ const Analytics = (): JSX.Element => {
 						name="dateRange"
 					/>
 				</form>
+				{isLoading ? (
+					<Loader />
+				) : (
+					<AnalyticsTable
+						activityLogs={activityLogs}
+						dateRange={dateRangeValue}
+					/>
+				)}
 			</section>
 		</PageLayout>
 	);