+++ apps/frontend/src/pages/analytics/libs/helpers/helpers.ts
@@ -1,1 +1,3 @@
+export { getAnalyticsColumns } from "./get-analytics-columns/get-analytics-columns.helper.jsx";
+export { getAnalyticsRows } from "./get-analytics-rows/get-analytics-rows.helper.js";
 export { getProjectOptions } from "./get-project-options.helper.js";