+++ packages/shared/src/modules/activity-logs/libs/enums/activity-log-validation-message.enum.ts
@@ -4,7 +4,9 @@ const ActivityLogValidationMessage = {
 	AUTHOR_NAME_REQUIRED: "Author name is required.",
 	COMMITS_NUMBER_REQUIRED: "Commits number is required.",
 	DATE_REQUIRED: "Date is required.",
+	END_DATE_REQUIRED: "Query end date is required.",
 	ITEMS_REQUIRED: "Activity log items cannot be empty.",
+	START_DATE_REQUIRED: "Query start date is required.",
 	USER_ID_REQUIRED: "User id is required.",
 } as const;