+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [1.29.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.28.1...backend-v1.29.0) (2024-09-20)
+
+
+### Features
+
+* add analytics table gf-327 ([#364](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/364)) ([dd7084a](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/dd7084abd8cf65b930897043a2ca837b1468745a))
+
+
+### Bug Fixes
+
+* Block access to project access management page when user hasn't either View or Manage All Projects permission gf-350 ([#414](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/414)) ([168101c](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/168101c05c7e20a38ff28bdca615db04dafaa4a8))
+
+## [1.28.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.28.0...backend-v1.28.1) (2024-09-20)
+
+
+### Bug Fixes
+
+* manage all projects permissions, edit, view, delete, create gf-411 ([#424](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/424)) ([3599a79](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/3599a79205f88979a958c150b73d287e54f0bfce))
+
 ## [1.28.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/backend-v1.27.0...backend-v1.28.0) (2024-09-19)